Move s3 cache into a service, out of controller [arclight]
